requirements.txt
playwright>=1.48.0
playwright-stealth>=1.0.6
python-dotenv>=1.0.1
